Inline SavePreferencesFactory to JabRefPreferences
# Conflicts:
#	src/main/java/org/jabref/gui/BasePanel.java
#	src/main/java/org/jabref/preferences/SavePreferencesFactory.java
