+++ src/main/java/org/jabref/gui/BasePanel.java
@@ -142,7 +142,6 @@ import org.jabref.model.entry.specialfields.SpecialFieldValue;
 import org.jabref.model.strings.StringUtil;
 import org.jabref.preferences.JabRefPreferences;
 import org.jabref.preferences.PreviewPreferences;
-import org.jabref.preferences.SavePreferencesFactory;
 
 import com.google.common.eventbus.Subscribe;
 import com.jgoodies.forms.builder.FormBuilder;
@@ -1093,7 +1092,7 @@ public class BasePanel extends JPanel implements ClipboardOwner {
         frame.block();
         final String SAVE_DATABASE = Localization.lang("Save library");
         try {
-            SavePreferences prefs = SavePreferencesFactory.loadForSaveFromPreferences(Globals.prefs)
+            SavePreferences prefs = JabRefPreferences.loadForSaveFromPreferences(Globals.prefs)
                     .withEncoding(enc)
                     .withSaveType(saveType);
             BibtexDatabaseWriter<SaveSession> databaseWriter = new BibtexDatabaseWriter<>(
